重构串口读取方法，将返回值类型从int32_t改为int64_t以支持更大数据量

+++ include/widget/io/Serial.h
@@ -303,7 +303,7 @@ namespace widget
 						   base::serial::StopBits stop_bits,
 						   base::serial::HardwareFlowControl hardware_flow_control) override;
 
-		virtual int32_t Read(base::Span const &span) override
+		virtual int64_t Read(base::Span const &span) override
 		{
 			return _received_stream.Read(span);
 		}
